requirements.txt
google-genai
python-dotenv
psycopg2
pgvector
streamlit
python-docx
PyPDF2
pandas
langchain
langchain-community
langchain-core
langchain-experimental
langchain-unstructured
unstructured
pytesseract
pdf2image
libmagic
docx2txt
tiktoken
markdown
